requirements.txt
# Requirements for Macro Signal Research Kit v3

# Core data science stack
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Database
prisma>=0.11.0
psycopg2-binary>=2.9.0

# Machine Learning
torch>=2.0.0
transformers>=4.30.0
optuna>=3.0.0
scikit-learn>=1.3.0

# Time series and forecasting
nixtla==0.6.6

# Data ingestion APIs
fredapi>=0.5.0
yfinance>=0.2.0
pandas-datareader>=0.10.0

# Visualization and export
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0
requests>=2.31.0

# Development tools
jupyter>=1.0.0
black>=23.0.0
pytest>=7.4.0

# Google AI for sentiment analysis
google-generativeai>=0.3.0

# Additional financial libraries
ta>=0.10.0  # Technical analysis
quantlib>=1.31  # Quantitative finance
